requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
groq==0.3.0
openai-whisper==20231117
python-dotenv==1.0.0
websockets==12.0
numpy==1.24.3
pydub==0.25.1
aiofiles==23.2.1
requests==2.31.0
torch==2.0.1
gtts==2.5.1
pyttsx3==2.90